requirements.txt
openai==1.66.3
tqdm
python-dotenv
pinecone==5.0.1
pinecone-datasets==1.0.2
pinecone-notebooks==0.1.1
pinecone-client==5.0.1
datasets==2.13.1